pyproject.toml
[tool.poetry]
name = "python-crawler"
version = "0.1.0"
description = "angels old lyrics crawler"
authors = ["Angel Reyna <rugile@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
pillow = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
